requirements.txt
shazamio>=0.7.0
aiohttp>=3.9
sounddevice>=0.4
soundfile>=0.12
pyyaml>=6
numpy>=1.24
